Fix no-unused-vars rule for object rest operator
If you try to use object rest with destructuring in order to pluck out
some properties from an object, the extension will throw an error for
unused variables.

E.g. if you do:
const {x, ...others} = params;
and plan to you use "others" variable only, ESLint will give an error
because x is unused. This is a false negative.

https://eslint.org/docs/rules/no-unused-vars#ignorerestsiblings

Change-Id: I80960a5e9e3fe02ca22652d7db9f3fa52cf434a2